refactor: :card_file_box: change database primary keys big Integer

+++ database/migrations/2024_07_22_152650_create_personnel_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('personnel', function (Blueprint $table) {
-            $table->integer('id')->unsigned();
+            $table->bigInteger('id')->unsigned();
             $table->integer('male_direct_re');
             $table->integer('female_direct_re');
             $table->integer('male_direct_part');
